refactor: Consolidate db dependency into graphql dir
