--- db/package.json
@@ -1,8 +1,0 @@
-{
-  "private": true,
-  "name": "@firstmeanseverything/db",
-  "version": "1.0.0",
-  "dependencies": {
-    "mongoose": "5.9.2"
-  }
-}
