+++ graphql/db/competition.js
